+++ src/lib_agnostic_baker/command_run.ml
@@ -155,7 +155,6 @@ let run_baker (module Plugin : Protocol_plugin_sig.S)
         force_apply_from_round;
         keep_alive;
         liquidity_baking_vote;
-        adaptive_issuance_vote;
         per_block_vote_file;
         extra_operations;
         dal_node_endpoint;
@@ -178,9 +177,6 @@ let run_baker (module Plugin : Protocol_plugin_sig.S)
          look into default locations. *)
       Per_block_vote_file.lookup_default_vote_file_path cctxt
     else Lwt.return per_block_vote_file
-  in
-  let*! () =
-    Events.warn_if_adaptive_issuance_vote_present ~adaptive_issuance_vote
   in
   (* We don't let the user run the baker without providing some
      option (CLI, file path, or file in default location) for