+++ src/lib_agnostic_baker/configuration.mli
@@ -25,7 +25,6 @@ val baker_args :
     * int option
     * bool
     * Per_block_votes.per_block_vote option
-    * Per_block_votes.per_block_vote option
     * string option
     * Uri.t option
     * Uri.t option
@@ -60,7 +59,6 @@ type t = {
   force_apply_from_round : int option;
   keep_alive : bool;
   liquidity_baking_vote : Per_block_votes.per_block_vote option;
-  adaptive_issuance_vote : Per_block_votes.per_block_vote option;
   per_block_vote_file : string option;
   extra_operations : Uri.t option;
   dal_node_endpoint : Uri.t option;
@@ -80,7 +78,6 @@ val create_config :
   * Q.t
   * int option
   * bool
-  * Per_block_votes.per_block_vote option
   * Per_block_votes.per_block_vote option
   * string option
   * Uri.t option