+++ src/proto_024_PtTALLiN/lib_delegate/baking_errors.ml
@@ -203,12 +203,8 @@ let () =
         ppf
         "@[The provided block vote file \"%s\" is a valid JSON file but its \
          content is unexpected. Expecting a JSON file containing \
-         '{\"liquidity_baking_toggle_vote\": value1, \
-         \"adaptive_issuance_vote\": value2}' or '{\"adaptive_issuance_vote\": \
-         value1, \"liquidity_baking_toggle_vote\": value2}', where value1 is \
-         one of \"on\", \"off\", or \"pass\" and value2 is one of \"on\", \
-         \"off\", or \"pass\", or '{\"liquidity_baking_toggle_vote\": value}' \
-         where value is one of \"on\", \"off\", or \"pass\".@]"
+         '{\"liquidity_baking_toggle_vote\": value}' where value is one of \
+         \"on\", \"off\", or \"pass\".@]"
         file_path)
     Data_encoding.(obj1 (req "file_path" string))
     (function
@@ -229,12 +225,8 @@ let () =
         ppf
         "@[In the provided block vote file \"%s\", the \
          \"liquidity_baking_toggle_vote\" field is missing. Expecting a JSON \
-         file containing '{\"liquidity_baking_toggle_vote\": value1, \
-         \"adaptive_issuance_vote\": value2}' or '{\"adaptive_issuance_vote\": \
-         value1, \"liquidity_baking_toggle_vote\": value2}', where value1 is \
-         one of \"on\", \"off\", or \"pass\" and value2 is one of \"on\", \
-         \"off\", or \"pass\", or '{\"liquidity_baking_toggle_vote\": value}' \
-         where value is one of \"on\", \"off\", or \"pass\".@]"
+         file containing '{\"liquidity_baking_toggle_vote\": value}' where \
+         value is one of \"on\", \"off\", or \"pass\".@]"
         file_path)
     Data_encoding.(obj1 (req "file_path" string))
     (function